+++ tests/Kiota.Builder.Tests/Plugins/PluginsGenerationServiceTests.cs
@@ -33,8 +33,11 @@ public sealed class PluginsGenerationServiceTests : IDisposable
         _httpClient.Dispose();
     }
 
-    [Fact]
-    public async Task GeneratesManifest()
+    [Theory]
+    [InlineData("client", "client")]
+    [InlineData("Budget Tracker", "BudgetTracker")]//drop the space
+    [InlineData("My-Super complex() %@#$& Name", "MySupercomplexName")]//drop the space and special characters
+    public async Task GeneratesManifest(string inputPluginName, string expectedPluginName)
     {
         var simpleDescriptionContent = @"openapi: 3.0.0
 info:
@@ -76,7 +79,7 @@ paths:
             OutputPath = outputDirectory,
             OpenAPIFilePath = "openapiPath",
             PluginTypes = [PluginType.APIPlugin, PluginType.APIManifest, PluginType.OpenAI],
-            ClientClassName = "client",
+            ClientClassName = inputPluginName,
             ApiRootUrl = "http://localhost/", //Kiota builder would set this for us
         };
         var (openAPIDocumentStream, _) = await openAPIDocumentDS.LoadStreamAsync(simpleDescriptionPath, generationConfiguration, null, false);
@@ -87,21 +90,22 @@ paths:
         var pluginsGenerationService = new PluginsGenerationService(openApiDocument, urlTreeNode, generationConfiguration, workingDirectory);
         await pluginsGenerationService.GenerateManifestAsync();
 
-        Assert.True(File.Exists(Path.Combine(outputDirectory, ManifestFileName)));
-        Assert.True(File.Exists(Path.Combine(outputDirectory, "client-apimanifest.json")));
+        Assert.True(File.Exists(Path.Combine(outputDirectory, $"{expectedPluginName.ToLower()}-apiplugin.json")));
+        Assert.True(File.Exists(Path.Combine(outputDirectory, $"{expectedPluginName.ToLower()}-apimanifest.json")));
         Assert.True(File.Exists(Path.Combine(outputDirectory, OpenAIPluginFileName)));
-        Assert.True(File.Exists(Path.Combine(outputDirectory, OpenApiFileName)));
+        Assert.True(File.Exists(Path.Combine(outputDirectory, $"{expectedPluginName.ToLower()}-openapi.yml")));
         Assert.True(File.Exists(Path.Combine(outputDirectory, AppManifestFileName)));
         Assert.True(File.Exists(Path.Combine(outputDirectory, "color.png")));
         Assert.True(File.Exists(Path.Combine(outputDirectory, "outline.png")));
 
         // Validate the v2 plugin
-        var manifestContent = await File.ReadAllTextAsync(Path.Combine(outputDirectory, ManifestFileName));
+        var manifestContent = await File.ReadAllTextAsync(Path.Combine(outputDirectory, $"{expectedPluginName.ToLower()}-apiplugin.json"));
         using var jsonDocument = JsonDocument.Parse(manifestContent);
         var resultingManifest = PluginManifestDocument.Load(jsonDocument.RootElement);
         Assert.NotNull(resultingManifest.Document);
-        Assert.Equal(OpenApiFileName, resultingManifest.Document.Runtimes.OfType<OpenApiRuntime>().First().Spec.Url);
+        Assert.Equal($"{expectedPluginName.ToLower()}-openapi.yml", resultingManifest.Document.Runtimes.OfType<OpenApiRuntime>().First().Spec.Url);
         Assert.Equal(2, resultingManifest.Document.Functions.Count);// all functions are generated despite missing operationIds
+        Assert.Equal(expectedPluginName, resultingManifest.Document.Namespace);// namespace is cleaned up.
         Assert.Empty(resultingManifest.Problems);// no problems are expected with names
 
         // Validate the v1 plugin
@@ -109,20 +113,20 @@ paths:
         using var v1JsonDocument = JsonDocument.Parse(v1ManifestContent);
         var v1Manifest = PluginManifestDocument.Load(v1JsonDocument.RootElement);
         Assert.NotNull(resultingManifest.Document);
-        Assert.Equal(OpenApiFileName, v1Manifest.Document.Api.URL);
+        Assert.Equal($"{expectedPluginName.ToLower()}-openapi.yml", v1Manifest.Document.Api.URL);
         Assert.Empty(v1Manifest.Problems);
 
         // Validate the manifest file
         var appManifestFile = await File.ReadAllTextAsync(Path.Combine(outputDirectory, AppManifestFileName));
         var appManifestModelObject = JsonSerializer.Deserialize(appManifestFile, PluginsGenerationService.AppManifestModelGenerationContext.AppManifestModel);
-        Assert.Equal("com.microsoft.kiota.plugin.client", appManifestModelObject.PackageName);
-        Assert.Equal("client", appManifestModelObject.Name.ShortName);
+        Assert.Equal($"com.microsoft.kiota.plugin.{expectedPluginName}", appManifestModelObject.PackageName);
+        Assert.Equal(expectedPluginName, appManifestModelObject.Name.ShortName);
         Assert.Equal("Microsoft Kiota.", appManifestModelObject.Developer.Name);
         Assert.Equal("color.png", appManifestModelObject.Icons.Color);
         Assert.NotNull(appManifestModelObject.CopilotExtensions.Plugins);
         Assert.Single(appManifestModelObject.CopilotExtensions.Plugins);
-        Assert.Equal("client", appManifestModelObject.CopilotExtensions.Plugins[0].Id);
-        Assert.Equal(ManifestFileName, appManifestModelObject.CopilotExtensions.Plugins[0].File);
+        Assert.Equal(expectedPluginName, appManifestModelObject.CopilotExtensions.Plugins[0].Id);
+        Assert.Equal($"{expectedPluginName.ToLower()}-apiplugin.json", appManifestModelObject.CopilotExtensions.Plugins[0].File);
     }
     private const string ManifestFileName = "client-apiplugin.json";
     private const string OpenAIPluginFileName = "openai-plugins.json";
@@ -687,6 +691,7 @@ paths:
   /test:
     get:
       description: description for test path
+      x-random-extension: true
       responses:
         '200':
           description: test
@@ -696,6 +701,7 @@ paths:
     get:
       description: description for test path with id
       operationId: test.WithId
+      x-openai-isConsequential: true
       parameters:
       - name: id
         in: path
@@ -768,7 +774,9 @@ components:
         Assert.Single(originalDocument.Extensions); // single unsupported extension at root
         Assert.Equal(2, originalDocument.Paths.Count); // document has only two paths
         Assert.Equal(2, originalDocument.Paths["/test"].Operations[OperationType.Get].Responses.Count); // 2 responses originally
+        Assert.Single(originalDocument.Paths["/test"].Operations[OperationType.Get].Extensions); // 1 UNsupported extension
         Assert.Equal(2, originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses.Count); // 2 responses originally
+        Assert.Single(originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Extensions); // 1 supported extension
 
         // Validate the output open api file
         var resultOpenApiFile = File.OpenRead(Path.Combine(outputDirectory, OpenApiFileName));
@@ -781,7 +789,9 @@ components:
         Assert.Equal(2, resultDocument.Paths.Count); // document has only two paths
         Assert.Single(resultDocument.Paths["/test"].Operations[OperationType.Get].Responses); // other responses are removed from the document
         Assert.NotEmpty(resultDocument.Paths["/test"].Operations[OperationType.Get].Responses["2XX"].Description); // response description string is not empty
+        Assert.Empty(resultDocument.Paths["/test"].Operations[OperationType.Get].Extensions); // NO UNsupported extension
         Assert.Single(resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses); // 2 responses originally
         Assert.NotEmpty(resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses["2XX"].Description);// response description string is not empty
+        Assert.Single(resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Extensions); // 1 supported extension still present in operation
     }
 }